Refactor: `GET /my-page/bookmarks` 북마크 리스트 조회 API 피드백 수정
