+++ src/main/java/com/ecolink/core/bookmark/service/BookmarkService.java
@@ -26,33 +26,33 @@ public class BookmarkService {
 	private final AvatarService avatarService;
 	private final StoreService storeService;
 
-	public boolean existsBookmark(Long avatarId, Long storeId) {
-		return bookmarkRepository.existsByAvatarAndStore(avatarId, storeId);
+	public boolean existsBookmark(Long storeId, Long avatarId) {
+		return bookmarkRepository.existsByAvatarAndStore(storeId, avatarId);
 	}
 
 	@Transactional
-	public void addBookmark(Long avatarId, Long storeId) {
-		Avatar avatar = avatarService.getById(avatarId);
+	public void addBookmark(Long storeId, Long avatarId) {
 		Store store = storeService.getById(storeId);
+		Avatar avatar = avatarService.getById(avatarId);
 
-		if (existsBookmark(avatarId, storeId)) {
+		if (existsBookmark(storeId, avatarId)) {
 			throw new BookmarkAlreadyExistsException(ErrorCode.BOOKMARK_ALREADY_EXISTS);
 		}
 
-		Bookmark bookmark = new Bookmark(avatar, store);
+		Bookmark bookmark = new Bookmark(store, avatar);
 		bookmarkRepository.save(bookmark);
 
 		store.addBookmarkCount();
 	}
 
-	public Bookmark getBookmark(Long avatarId, Long storeId) {
-		return bookmarkRepository.findBookmarkByAvatarIdAndStoreId(avatarId, storeId)
+	public Bookmark getBookmark(Long storeId, Long avatarId) {
+		return bookmarkRepository.findBookmarkByAvatarIdAndStoreId(storeId, avatarId)
 			.orElseThrow(() -> new BookmarkNotFoundException(ErrorCode.BOOKMARK_NOT_FOUND));
 	}
 
 	@Transactional
-	public void deleteBookmark(Long avatarId, Long storeId) {
-		Bookmark bookmark = getBookmark(avatarId, storeId);
+	public void deleteBookmark(Long storeId, Long avatarId) {
+		Bookmark bookmark = getBookmark(storeId, avatarId);
 		Store store = bookmark.getStore();
 
 		bookmarkRepository.delete(bookmark);
